Added projects, buttons and slider changes
13 projects added to the projects slider page, in the slider home the projects has been changed to the newest 6.
Added 'solicitar cotizacion' buttons in 'hero landing', 'testimonios' and 'nuestros servicios' sections in home as well as it's corresponding styles.
Changed whatsapp fixed 'cotiza tu proyecto' button to svg, added the text and hover styles.
Projects slider has a temporary fix where when the user click to change to the next image, it scrolls the page to align the section to avoid unintentionally page jumps.